pal_finder: new 'pal_filter.py' from Graeme Fox replaces 'pal_finder_filter_and_assembly.py'.
